Add Kotlin API dump plugin (#184)
Unfortunately this plugin does not actually validate that changes are compatible so reviewers must still remain vigilant to changes to this file. Obviously we're pre-1.0 so we can tolerate some breaking changes, and this will help make us aware of it.

Long term we deserve something better than this, but this is good enough for now.